add aarch64, arm support
