+++ patcherex/arch/i386/utils.py
@@ -1,0 +1,1 @@
+from patcherex.utils import *
